+++ client/src/App.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import 'antd/dist/antd.css';
 
 import { BrowserRouter as Router } from 'react-router-dom';
